wip - rework of user interface
